requirements.txt
gradio==4.15.0
python-dotenv==1.0.0
openai==1.3.0
reportlab==4.0.4
